requirements.txt
praw 
google-api-python-client 
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
speechrecognition
pyaudio